+++ src/main/java/projeto/gabriel/projeto/repository/CompraRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import projeto.gabriel.projeto.model.Compra;
 
-public interface CompraRepository extends CrudRepository<Compra, Long> { }
+public interface CompraRepository extends CrudRepository<Compra, Integer> { }
