+++ src/main/java/projeto/gabriel/projeto/repository/ProdutoCompraRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import projeto.gabriel.projeto.model.ProdutoCompra;
 
-public interface ProdutoCompraRepository extends CrudRepository<ProdutoCompra, Long> { }
+public interface ProdutoCompraRepository extends CrudRepository<ProdutoCompra, Integer> { }
