+++ src/main/java/projeto/gabriel/projeto/repository/ProdutoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import projeto.gabriel.projeto.model.Produto;
 
-public interface ProdutoRepository extends CrudRepository<Produto, Long> { }
+public interface ProdutoRepository extends CrudRepository<Produto, Integer> { }
