+++ src/main/java/projeto/gabriel/projeto/repository/UsuarioRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository;
 
 import projeto.gabriel.projeto.model.Usuario;
 
-public interface UsuarioRepository extends CrudRepository<Usuario, Long> { }
+public interface UsuarioRepository extends CrudRepository<Usuario, Integer> { }
